Move reservation-confirmation to shared
Refs #499.
