+++ app/pages/resource/ResourcePage.js
@@ -7,10 +7,10 @@ import { bindActionCreators } from 'redux';
 import { fetchResource } from 'actions/resourceActions';
 import PageWrapper from 'pages/PageWrapper';
 import NotFoundPage from 'pages/not-found/NotFoundPage';
+import ReservationConfirmation from 'shared/reservation-confirmation';
 import { injectT } from 'i18n';
 import { getDateStartAndEndTimes } from 'utils/timeUtils';
 import ReservationCalendar from './reservation-calendar';
-import ReservationConfirmation from './reservation-confirmation';
 import ReservationInfo from './reservation-info';
 import ResourceInfo from './resource-info';
 import resourcePageSelector from './resourcePageSelector';
