+++ app/pages/resource/resource-page.less
@@ -1,5 +1,4 @@
 @import './reservation-calendar/reservation-calendar';
 @import './reservation-calendar/time-slots/time-slots';
-@import './reservation-confirmation/confirm-reservation-modal';
 @import './resource-info/image-carousel';
 @import './resource-info/resource-info';
