+++ app/shared/reservation-confirmation/confirm-reservation-modal.less
