+++ app/shared/reservation-confirmation/index.js
