+++ app/shared/reservation-confirmation/reservationConfirmationSelector.js
