+++ app/shared/shared.less
@@ -9,6 +9,7 @@
 @import './modals/modals';
 @import './navbar/navbar';
 @import './notifications/notifications';
+@import './reservation-confirmation/confirm-reservation-modal';
 @import './resource-icons/resource-icons';
 @import './resource-list/resource-list';
 @import './test-site-message/test-site-message';
